build(generate): some more refactoring & fixes
