+++ deno.lock
@@ -1647,6 +1647,7 @@
     "https://esm.sh/v134/type-fest@4.8.1/denonext/source/union-to-intersection.d.js": "4dda98575465ccb71439587d666e602980113bf8d01bb42067ca6ffeb4fae0e3",
     "https://esm.sh/v134/type-fest@4.8.1/source/simplify.d": "75d5f739975152accd8282ecb89818cc4644e1ee2c73b13055de5e0ad6349f2e",
     "https://esm.sh/v134/type-fest@4.8.1/source/union-to-intersection.d": "7be1a6951fbb06a27c1ad9820a5af2571c77adb65d8fb818ef7c7053e6ed14fa",
+    "https://raw.githubusercontent.com/SchemaStore/schemastore/404ada34125022391ae5a4484d490f3d769a6555/src/schemas/json/github-issue-forms.json": "b34a50fecd7797e6c871afbadaecd4edffd62df15b0c728ed6b67861798341d4",
     "https://raw.githubusercontent.com/catppuccin/catppuccin/91d6b5433730103c504dcf43583b594e418ee7c1/resources/ports.schema.json": "f1bfc1f4c68af258cd2f1a254d70069b7d254ab0cfe298b9b52bcd78299d8e40"
   }
 }
